Add backend module and pizza class
Created a few basic endpoints to use such as add_to_cart, remove_from_cart, and get_cart. They don't work yet, but they can start be used to the front end.

Also created a Pizza class as a consolidated place to store variables about a pizza, such as size, crust, and toppings. This makes it much simpler to pass a pizza as a parameter.